Refactor data fields to own module
