+++ module/data/actor/templates/sentient.mjs
@@ -1,7 +1,7 @@
 import CommonTemplate from "./common.mjs";
-import { MappingField } from "../../fields.mjs";
 import MovementFields from "./movement.mjs";
 import ED4E from "../../../config.mjs";
+import MappingField from "../../fields/mapping-field.mjs";
 
 /**
  * A template for all actors that represent sentient beings and have such stats.
