+++ module/data/fields/datetime-field.mjs
@@ -1,0 +1,3 @@
+export default class DatetimeField extends foundry.data.fields.DataField {
+  // TODO: implement
+}
