+++ module/data/fields/identifier-field.mjs
@@ -1,0 +1,13 @@
+/**
+ * Special case StringField that includes automatic validation for identifiers.
+ */
+export default class IdentifierField extends foundry.data.fields.StringField {
+  /**
+   * @override
+   */
+  _validateType( value ) {
+    if ( !ed4e.utils.validators.isValidIdentifier( value ) ) {
+      throw new Error( game.i18n.localize( "ED.Errors.IdentifierError" ) );
+    }
+  }
+}
