+++ module/data/fields/mapping-field.mjs
@@ -1,103 +1,3 @@
-import ED4E from "../config.mjs";
-import { slugify, validateEdid } from "../utils.mjs";
-import { getLocalizeKey } from "./abstract.mjs";
-
-/**
- * Taken from the ({@link https://gitlab.com/peginc/swade/-/wikis/Savage-Worlds-ID|SWADE system}).
- * A special case string field that represents a strictly slugged string.
- */
-export class EdIdField extends foundry.data.fields.StringField {
-
-  /** @inheritdoc */
-  static get _defaults() {
-    return foundry.utils.mergeObject( super._defaults, {
-      initial:  ED4E.reserved_edid.DEFAULT,
-      blank:    false,
-      required: true,
-      label:    getLocalizeKey( "Item", false, "edid" ),
-      hints:    getLocalizeKey( "Item", true, "edid" ),
-    } );
-  }
-
-  /**
-   * @override
-   */
-  clean( value, options ) {
-    const slug = slugify( value );
-    // only return slug if non-empty so empty slugs will be shown as errors
-    return super.clean( slug ? slug : value, options );
-  }
-
-  /**
-   * @override
-   */
-  _validateType( value, _ ) {
-    return validateEdid( value );
-  }
-}
-
-/*
-* Field implementations are taken from the [DnD5e system]{@link https://github.com/foundryvtt/dnd5e}
-*/
-
-/**
- * Special case StringField that includes automatic validation for identifiers.
- */
-export class IdentifierField extends foundry.data.fields.StringField {
-  /**
-   * @override
-   */
-  _validateType( value ) {
-    if ( !ed4e.utils.validators.isValidIdentifier( value ) ) {
-      throw new Error( game.i18n.localize( "ED.Errors.IdentifierError" ) );
-    }
-  }
-}
-
-/* -------------------------------------------- */
-
-/**
- * @typedef {StringFieldOptions} FormulaFieldOptions
- * @property {boolean} [deterministic=false]  Is this formula not allowed to have dice values?
- */
-
-/**
- * Special case StringField which represents a formula.
- * @param {FormulaFieldOptions} [options={}]  Options which configure the behavior of the field.
- * @property {boolean} [deterministic=false]  Is this formula not allowed to have dice values?
- */
-export class FormulaField extends foundry.data.fields.StringField {
-
-  /** @inheritdoc */
-  static get _defaults() {
-    return foundry.utils.mergeObject( super._defaults, {
-      deterministic: false
-    } );
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  _validateType( value ) {
-    if ( this.options.deterministic ) {
-      const roll = new Roll( value );
-      if ( !roll.isDeterministic ) throw new Error( "must not contain dice terms" );
-      Roll.safeEval( roll.formula );
-    }
-    else Roll.validate( value );
-    super._validateType( value );
-  }
-}
-
-/* -------------------------------------------- */
-
-export class DatetimeField extends foundry.data.fields.DataField {
-  // TODO: implement
-}
-
-/* -------------------------------------------- */
-
-
 /**
  * @callback MappingFieldInitialValueBuilder
  * @param {string} key       The key within the object where this new value is being generated.
@@ -124,7 +24,7 @@ export class DatetimeField extends foundry.data.fields.DataField {
  * @property {boolean} [initialKeysOnly=false]  Should the keys in the initialized data be limited to the keys provided
  *                                              by `options.initialKeys`?
  */
-export class MappingField extends foundry.data.fields.ObjectField {
+export default class MappingField extends foundry.data.fields.ObjectField {
   constructor( model, options ) {
     if ( !( model instanceof foundry.data.fields.DataField ) ) {
       throw new Error( "MappingField must have a DataField as its contained element" );
@@ -136,45 +36,32 @@ export class MappingField extends foundry.data.fields.ObjectField {
      * @type {DataField}
      */
     this.model = model;
+    model.parent = this;
   }
 
   /* -------------------------------------------- */
 
-  /**
-   * @inheritDoc
-   */
+  /** @inheritDoc */
   static get _defaults() {
     return foundry.utils.mergeObject( super._defaults, {
       initialKeys:     null,
       initialValue:    null,
-      initialKeysOnly: false,
+      initialKeysOnly: false
     } );
   }
 
   /* -------------------------------------------- */
 
-  /**
-   * @inheritDoc
-   */
+  /** @inheritDoc */
   _cleanType( value, options ) {
-    Object.entries( value ).forEach(
-      // disable eslint for the rule `no-return-assign`
-      // this would apply in the following line, since arrow functions have an implicit `return`
-      // it is, however, good practice to write it in that short concise manner
-      // (the irony of the fact that this comment now make the code more unreadable than just using curly braces is not lost on me ;) )
-      // see https://github.com/eslint/eslint/issues/5150
-      /* eslint-disable */
-      ([k, v]) => (value[k] = this.model.clean(v, options)),
-      /* eslint-enable */
-    );
+    // eslint-disable-next-line no-return-assign
+    Object.entries( value ).forEach( ( [ k, v ] ) => value[k] = this.model.clean( v, options ) );
     return value;
   }
 
   /* -------------------------------------------- */
 
-  /**
-   * @inheritDoc
-   */
+  /** @inheritDoc */
   getInitialValue( data ) {
     let keys = this.initialKeys;
     const initial = super.getInitialValue( data );
@@ -201,16 +88,14 @@ export class MappingField extends foundry.data.fields.ObjectField {
 
   /* -------------------------------------------- */
 
-  /**
-   * @override
-   */
+  /** @override */
   _validateType( value, options = {} ) {
     if ( foundry.utils.getType( value ) !== "Object" ) throw new Error( "Must be an Object" );
     const errors = this._validateValues( value, options );
     if ( !foundry.utils.isEmpty( errors ) ) {
       throw new foundry.data.validation.DataModelValidationError(
         Object.entries( errors ).map(
-          ( [ k, v ] ) => `\n${k}: ${errors[k].toString()}`
+          ( [ k, _ ] ) => `\n${ k }: ${ errors[k].toString() }`
         ).join( "" )
       );
     }
@@ -235,12 +120,10 @@ export class MappingField extends foundry.data.fields.ObjectField {
 
   /* -------------------------------------------- */
 
-  /**
-   * @override
-   */
+  /** @override */
   initialize( value, model, options = {} ) {
     if ( !value ) return value;
-    const obj ={};
+    const obj = {};
     const initialKeys = ( this.initialKeys instanceof Array ) ? this.initialKeys : Object.keys( this.initialKeys ?? {} );
     const keys = this.initialKeysOnly ? initialKeys : Object.keys( value );
     for ( const key of keys ) {
@@ -252,16 +135,12 @@ export class MappingField extends foundry.data.fields.ObjectField {
 
   /* -------------------------------------------- */
 
-  /**
-   * @inheritDoc
-   */
+  /** @inheritDoc */
   _getField( path ) {
-    if ( path.length === 0 ) {
-      return this;
-    } else if ( path.length === 1 ) {
-      return this.model;
-    }
+    if ( path.length === 0 ) return this;
+    else if ( path.length === 1 ) return this.model;
     path.shift();
     return this.model._getField( path );
   }
+
 }
